backend/requirements.txt
flask==3.0.0
flask-socketio==5.3.6
python-socketio==5.10.0
python-engineio==4.9.0
bidict==0.23.1
typing-extensions==4.14.0
annotated-types==0.7.0
pydantic==2.11.5
pydantic-core==2.33.2
typing-inspection==0.4.1
anyio==4.9.0
certifi==2025.4.26
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
sniffio==1.3.1
distro==1.9.0
audioop-lts==0.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
attrs==25.3.0
frozenlist==1.6.2
multidict==6.4.4
propcache==0.3.1
yarl==1.20.0
iterators==0.2.0
dnspython==2.7.0
eventlet==0.40.0
greenlet==3.2.2
openai==1.3.7
cartesia==2.0.3
pydub>=0.25.1
python-dotenv==1.0.0
requests==2.31.0
# SpeechRecognition>=3.10.0
webrtcvad>=2.0.10
websocket-client>=1.8.0
werkzeug==3.1.3
click==8.2.1
blinker==1.9.0
jinja2==3.1.6
itsdangerous==2.2.0
ruff==0.11.12